+++ bytebankHerdado/src/bytebankHerdado/Gerente.java
@@ -1,0 +1,23 @@
+package bytebankHerdado;
+
+//Gerente é um Funcionário ou Gerente herda da classe Funcionário
+public class Gerente extends Funcionario {
+
+	private int senha;
+	
+	public void setSenha(int senha) {
+		this.senha = senha;
+	}
+	
+	public boolean autentica(int senha) {
+		if (this.senha == senha) {
+			return true;
+		} else {
+			return false;
+		}
+	}
+	
+//	public double getBonificacao() {
+//		return this.salario;
+//	}	
+}
